Remove smappee integration and rename smappee_official
