+++ .coveragerc
@@ -726,11 +726,10 @@ omit =
     homeassistant/components/sinch/*
     homeassistant/components/slide/*
     homeassistant/components/sma/sensor.py
-    homeassistant/components/smappee/*
-    homeassistant/components/smappee_official/__init__.py
-    homeassistant/components/smappee_official/api.py
-    homeassistant/components/smappee_official/sensor.py
-    homeassistant/components/smappee_official/switch.py
+    homeassistant/components/smappee/__init__.py
+    homeassistant/components/smappee/api.py
+    homeassistant/components/smappee/sensor.py
+    homeassistant/components/smappee/switch.py
     homeassistant/components/smarty/*
     homeassistant/components/smarthab/*
     homeassistant/components/sms/*
