+++ homeassistant/components/smappee/api.py
