+++ homeassistant/components/smappee/const.py
