+++ homeassistant/components/smappee/strings.json
