+++ homeassistant/components/smappee/translations/en.json
