--- homeassistant/components/smappee_official/manifest.json
@@ -1,13 +1,0 @@
-{
-  "domain": "smappee_official",
-  "name": "Smappee Official",
-  "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/smappee_official",
-  "dependencies": ["http"],
-  "requirements": [
-    "pysmappee==0.1.0"
-  ],
-  "codeowners": [
-    "@bsmappee"
-  ]
-}
