+++ requirements_test_all.txt
@@ -701,7 +701,7 @@ pysignalclirestapi==0.3.4
 # homeassistant.components.sma
 pysma==0.3.5
 
-# homeassistant.components.smappee_official
+# homeassistant.components.smappee
 pysmappee==0.1.0
 
 # homeassistant.components.smartthings
