+++ tests/components/smappee/__init__.py
