+++ tests/components/smappee/test_config_flow.py
@@ -1,10 +1,6 @@
 """Test the Smappee Official config flow."""
 from homeassistant import config_entries, data_entry_flow, setup
-from homeassistant.components.smappee_official.const import (
-    AUTHORIZE_URL,
-    DOMAIN,
-    TOKEN_URL,
-)
+from homeassistant.components.smappee.const import AUTHORIZE_URL, DOMAIN, TOKEN_URL
 from homeassistant.const import CONF_CLIENT_ID, CONF_CLIENT_SECRET
 from homeassistant.helpers import config_entry_oauth2_flow
